APPS-2568 fix warning

+++ web/src/main/java/org/geogebra/web/html5/multiuser/TooltipChip.java
@@ -33,6 +33,6 @@ public class TooltipChip extends Label {
 		getElement().removeClassName("invisible");
 		Style style = getElement().getStyle();
 		style.setLeft(x + LEFT_MARGIN, Style.Unit.PX);
-		style.setTop(y - (double) (getOffsetHeight() / 2), Style.Unit.PX);
+		style.setTop(y - (getOffsetHeight() / 2d), Style.Unit.PX);
 	}
 }
